astrogrid/security/test/java/org/astrogrid/security/SecurityGuardTest.java
package org.astrogrid.security;


import java.util.Set;
import javax.security.auth.Subject;
import junit.framework.TestCase;





/**
 * Unit tests for the {@link SecurityGuard class}.
 * JUnit is used.
 *
 * The SUT is a Java bean with properties but no
 * business logic. Each test case tests the setter
 * and getter for one property, assuming that there
 * are no internal linkagaes between properties.
 *
 * @author Guy Rixon

 */

public class SecurityGuardTest extends TestCase {



  /**
   * Constructs the test suite.
   */

  public SecurityGuardTest () {}



  /**
   * Tests the ssoUsername property.
   */

  public void testSsoUsername () throws Exception {
    SecurityGuard g = new SecurityGuard();
    assertNull("Null SSO user-name after construction", g.getSsoUsername());
    g.setSsoUsername("fred");
    assertEquals("fred", g.getSsoUsername());
  }

  /**
   * Tests the ssoPassword property.
   */

  public void testSsoPassword () throws Exception {
    SecurityGuard g = new SecurityGuard();
    assertNull("Null SSO password after construction", g.getSsoPassword());
    g.setSsoPassword("secret");
    assertEquals("Password matches", "secret", g.getSsoPassword());
  }

}